Grid and Node components with  everything necessary  for the algorithms to find ther path
Opted for a class component for Nodes as the ref interface for addng fields to help the algorithms is nicer than useImperativeHandle

The grid will resize automatically. Node refs are stored in zustand in a map. These refs will be passed to the algorithms. Have implemented a throttle on resize for performance